Extracted functions for checking the user's rights to the helper file
